update user meta from heartbeat request
